src/main/java/io/hhplus/tdd/point/PointController.java
package io.hhplus.tdd.point;

import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/point")
public class PointController {

    private static final Logger log = LoggerFactory.getLogger(PointController.class);

    private final PointService pointService;

  public PointController(PointService pointService) {
    this.pointService = pointService;
  }

  /**
     * TODO - 특정 유저의 포인트를 조회하는 기능을 작성해주세요.
     */
    @GetMapping("{id}")
    public ResponseEntity<UserPoint> point(
            @PathVariable long id
    ) {
      log.info("point id: {}", id);
      UserPoint userPoint = pointService.point(id);
      return ResponseEntity.ok(userPoint);
    }

    /**
     * TODO - 특정 유저의 포인트 충전/이용 내역을 조회하는 기능을 작성해주세요.
     */
    @GetMapping("{id}/histories")
    public ResponseEntity<List<PointHistory>> history(
            @PathVariable long id
    ) {
      log.info("history id: {}", id);
      List<PointHistory> history = pointService.history(id);
        return ResponseEntity.ok(history);
    }

    /**
     * TODO - 특정 유저의 포인트를 충전하는 기능을 작성해주세요.
     */
    @PatchMapping("{id}/charge")
    public ResponseEntity<UserPoint> charge(
            @PathVariable long id,
            @RequestBody long amount
    ) {
      log.info("charge id: {}, amount: {}", id, amount);
      UserPoint charged = pointService.charge(id, amount);
      return ResponseEntity.ok(charged);
    }

    /**
     * TODO - 특정 유저의 포인트를 사용하는 기능을 작성해주세요.
     */
    @PatchMapping("{id}/use")
    public ResponseEntity<UserPoint> use(
            @PathVariable long id,
            @RequestBody long amount
    ) {
      log.info("use id: {}, amount: {}", id, amount);
      UserPoint use = pointService.use(id, amount);
      return ResponseEntity.ok(use);
    }
}
